# Conflicts:
#	cmd/loki-canary/main.go
#	pkg/canary/writer/writer.go

+++ integration/cluster/cluster.go
@@ -1,6 +1,7 @@
 package cluster
 
 import (
+	"bytes"
 	"context"
 	"errors"
 	"flag"
@@ -9,16 +10,26 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
+	"path/filepath"
 	"strings"
 	"sync"
 	"text/template"
 	"time"
 
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/multierror"
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/common/model"
+	"gopkg.in/yaml.v2"
+
+	"github.com/grafana/loki/integration/util"
 
 	"github.com/grafana/loki/pkg/loki"
+	"github.com/grafana/loki/pkg/storage/config"
 	"github.com/grafana/loki/pkg/util/cfg"
+	util_log "github.com/grafana/loki/pkg/util/log"
+	"github.com/grafana/loki/pkg/validation"
 )
 
 var (
@@ -30,6 +41,9 @@ auth_enabled: true
 server:
   http_listen_port: 0
   grpc_listen_port: 0
+  grpc_server_max_recv_msg_size: 110485813
+  grpc_server_max_send_msg_size: 110485813
+
 
 common:
   path_prefix: {{.dataPath}}
@@ -43,25 +57,27 @@ common:
     kvstore:
       store: inmemory
 
+limits_config:
+  per_stream_rate_limit: 50MB
+  per_stream_rate_limit_burst: 50MB
+  ingestion_rate_mb: 50
+  ingestion_burst_size_mb: 50
+  reject_old_samples: false
+
 storage_config:
+  named_stores:
+    filesystem:
+      store-1:
+        directory: {{.sharedDataPath}}/fs-store-1
   boltdb_shipper:
-    shared_store: filesystem
     active_index_directory: {{.dataPath}}/index
     cache_location: {{.dataPath}}/boltdb-cache
-
-schema_config:
-  configs:
-    - from: 2020-10-24
-      store: boltdb-shipper
-      object_store: filesystem
-      schema: v11
-      index:
-        prefix: index_
-        period: 24h
+  tsdb_shipper:
+    active_index_directory: {{.dataPath}}/tsdb-index
+    cache_location: {{.dataPath}}/tsdb-cache
 
 compactor:
   working_directory: {{.dataPath}}/retention
-  shared_store: filesystem
   retention_enabled: true
 
 analytics:
@@ -71,43 +87,25 @@ ingester:
   lifecycler:
     min_ready_duration: 0s
 
-{{if .remoteWriteUrls}}
+querier:
+  multi_tenant_queries_enabled: true
+
+query_scheduler:
+  max_outstanding_requests_per_tenant: 2048
+
 ruler:
-  wal:
-    dir: {{.rulerWALPath}}
-  storage:
-    type: local
-    local:
-      directory: {{.rulesPath}}
-  rule_path: {{.sharedDataPath}}/rule
   enable_api: true
   ring:
     kvstore:
       store: inmemory
-  remote_write:
-    enabled: true
-    clients:
-      remote_client1:
-        url: {{index .remoteWriteUrls 0}}/api/v1/write
-      remote_client2:
-        url: {{index .remoteWriteUrls 1}}/api/v1/write
-{{end}}
+  wal:
+    dir: {{.sharedDataPath}}/ruler-wal
+  storage:
+    type: local
+    local:
+      directory: {{.sharedDataPath}}/rules
+  rule_path: {{.sharedDataPath}}/prom-rule
 `))
-
-	rulesConfig = `
-groups:
-- name: always-firing
-  interval: 1s
-  rules:
-  - alert: fire
-    expr: |
-      1 > 0
-    for: 0m
-    labels:
-      severity: warning
-    annotations:
-      summary: test
-`
 )
 
 func wrapRegistry() {
@@ -140,21 +138,43 @@ func (w *wrappedRegisterer) MustRegister(collectors ...prometheus.Collector) {
 }
 
 type Cluster struct {
-	sharedPath string
-	components []*Component
-	waitGroup  sync.WaitGroup
+	sharedPath    string
+	components    []*Component
+	waitGroup     sync.WaitGroup
+	initedAt      model.Time
+	periodCfgs    []string
+	overridesFile string
 }
 
-func New() *Cluster {
+func New(logLevel level.Value, opts ...func(*Cluster)) *Cluster {
+	if logLevel != nil {
+		util_log.Logger = level.NewFilter(log.NewLogfmtLogger(os.Stderr), level.Allow(logLevel))
+	}
+
 	wrapRegistry()
 	sharedPath, err := os.MkdirTemp("", "loki-shared-data")
 	if err != nil {
 		panic(err.Error())
 	}
 
-	return &Cluster{
-		sharedPath: sharedPath,
+	overridesFile := filepath.Join(sharedPath, "loki-overrides.yaml")
+
+	err = os.WriteFile(filepath.Join(sharedPath, "loki-overrides.yaml"), []byte(`overrides:`), 0777)
+	if err != nil {
+		panic(fmt.Errorf("error creating overrides file: %w", err))
 	}
+
+	cluster := &Cluster{
+		sharedPath:    sharedPath,
+		initedAt:      model.Now(),
+		overridesFile: overridesFile,
+	}
+
+	for _, opt := range opts {
+		opt(cluster)
+	}
+
+	return cluster
 }
 
 func (c *Cluster) Run() error {
@@ -169,7 +189,20 @@ func (c *Cluster) Run() error {
 	}
 	return nil
 }
+
+func (c *Cluster) Restart() error {
+	if err := c.stop(false); err != nil {
+		return err
+	}
+
+	return c.Run()
+}
+
 func (c *Cluster) Cleanup() error {
+	return c.stop(true)
+}
+
+func (c *Cluster) stop(cleanupFiles bool) error {
 	_, cancelFunc := context.WithTimeout(context.Background(), time.Second*3)
 	defer cancelFunc()
 
@@ -195,12 +228,14 @@ func (c *Cluster) Cleanup() error {
 	// wait for all process to close
 	c.waitGroup.Wait()
 
-	// cleanup dirs/files
-	for _, d := range dirs {
-		errs.Add(os.RemoveAll(d))
-	}
-	for _, f := range files {
-		errs.Add(os.Remove(f))
+	if cleanupFiles {
+		// cleanup dirs/files
+		for _, d := range dirs {
+			errs.Add(os.RemoveAll(d))
+		}
+		for _, f := range files {
+			errs.Add(os.Remove(f))
+		}
 	}
 
 	return errs.Err()
@@ -208,10 +243,11 @@ func (c *Cluster) Cleanup() error {
 
 func (c *Cluster) AddComponent(name string, flags ...string) *Component {
 	component := &Component{
-		name:    name,
-		cluster: c,
-		flags:   flags,
-		running: false,
+		name:          name,
+		cluster:       c,
+		flags:         flags,
+		running:       false,
+		overridesFile: c.overridesFile,
 	}
 
 	c.components = append(c.components, component)
@@ -224,15 +260,24 @@ type Component struct {
 	cluster *Cluster
 	flags   []string
 
-	configFile   string
-	dataPath     string
-	rulerWALPath string
-	rulesPath    string
-	RulesTenant  string
+	configFile    string
+	extraConfigs  []string
+	overridesFile string
+	dataPath      string
 
 	running bool
+	wg      sync.WaitGroup
+}
 
-	RemoteWriteUrls []string
+// ClusterSharedPath returns the path to the shared directory between all components in the cluster.
+// This path will be removed once the cluster is stopped.
+func (c *Component) ClusterSharedPath() string {
+	return c.cluster.sharedPath
+}
+
+// component should be restarted if it's already running for the new flags to take effect
+func (c *Component) AddFlags(flags ...string) {
+	c.flags = append(c.flags, flags...)
 }
 
 func (c *Component) HTTPURL() string {
@@ -241,6 +286,14 @@ func (c *Component) HTTPURL() string {
 
 func (c *Component) GRPCURL() string {
 	return fmt.Sprintf("localhost:%s", port(c.loki.Server.GRPCListenAddr().String()))
+}
+
+func (c *Component) WithExtraConfig(cfg string) {
+	if c.running {
+		panic("cannot set extra config after component is running")
+	}
+
+	c.extraConfigs = append(c.extraConfigs, cfg)
 }
 
 func port(addr string) string {
@@ -261,44 +314,12 @@ func (c *Component) writeConfig() error {
 		return fmt.Errorf("error creating data path: %w", err)
 	}
 
-	if len(c.RemoteWriteUrls) > 0 {
-		c.rulesPath, err = os.MkdirTemp(c.cluster.sharedPath, "rules")
-		if err != nil {
-			return fmt.Errorf("error creating rules path: %w", err)
-		}
-
-		fakeDir, err := os.MkdirTemp(c.rulesPath, "fake")
-		if err != nil {
-			return fmt.Errorf("error creating rules/fake path: %w", err)
-		}
-
-		s := strings.Split(fakeDir, "/")
-		c.RulesTenant = s[len(s)-1]
-
-		c.rulerWALPath, err = os.MkdirTemp(c.cluster.sharedPath, "ruler-wal")
-		if err != nil {
-			return fmt.Errorf("error creating ruler-wal path: %w", err)
-		}
-
-		rulesConfigFile, err := os.CreateTemp(fakeDir, "rules*.yaml")
-		if err != nil {
-			return fmt.Errorf("error creating rules config file: %w", err)
-		}
-
-		if _, err = rulesConfigFile.Write([]byte(rulesConfig)); err != nil {
-			return fmt.Errorf("error writing to rules config file: %w", err)
-		}
-
-		rulesConfigFile.Close()
+	mergedConfig, err := c.MergedConfig()
+	if err != nil {
+		return fmt.Errorf("error getting merged config: %w", err)
 	}
 
-	if err := configTemplate.Execute(configFile, map[string]interface{}{
-		"dataPath":        c.dataPath,
-		"sharedDataPath":  c.cluster.sharedPath,
-		"remoteWriteUrls": c.RemoteWriteUrls,
-		"rulesPath":       c.rulesPath,
-		"rulerWALPath":    c.rulerWALPath,
-	}); err != nil {
+	if err := os.WriteFile(configFile.Name(), mergedConfig, 0644); err != nil {
 		return fmt.Errorf("error writing config file: %w", err)
 	}
 
@@ -307,6 +328,52 @@ func (c *Component) writeConfig() error {
 	}
 	c.configFile = configFile.Name()
 	return nil
+}
+
+// MergedConfig merges the base config template with any additional config that has been provided
+func (c *Component) MergedConfig() ([]byte, error) {
+	var sb bytes.Buffer
+
+	periodStart := config.DayTime{Time: c.cluster.initedAt.Add(-24 * time.Hour)}
+	additionalPeriodStart := config.DayTime{Time: c.cluster.initedAt.Add(-7 * 24 * time.Hour)}
+
+	if err := configTemplate.Execute(&sb, map[string]interface{}{
+		"dataPath":       c.dataPath,
+		"sharedDataPath": c.cluster.sharedPath,
+	}); err != nil {
+		return nil, fmt.Errorf("error writing config file: %w", err)
+	}
+
+	merger := util.NewYAMLMerger()
+	merger.AddFragment(sb.Bytes())
+
+	// default to using boltdb index
+	if len(c.cluster.periodCfgs) == 0 {
+		c.cluster.periodCfgs = []string{boltDBShipperSchemaConfigTemplate}
+	}
+
+	for _, periodCfg := range c.cluster.periodCfgs {
+		var buf bytes.Buffer
+		if err := template.Must(template.New("schema").Parse(periodCfg)).
+			Execute(&buf, map[string]interface{}{
+				"curPeriodStart":        periodStart.String(),
+				"additionalPeriodStart": additionalPeriodStart.String(),
+			}); err != nil {
+			return nil, errors.New("error building schema_config")
+		}
+		merger.AddFragment(buf.Bytes())
+	}
+
+	for _, extra := range c.extraConfigs {
+		merger.AddFragment([]byte(extra))
+	}
+
+	merged, err := merger.Merge()
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal merged config to YAML: %w", err)
+	}
+
+	return merged, nil
 }
 
 func (c *Component) run() error {
@@ -324,6 +391,8 @@ func (c *Component) run() error {
 		c.flags,
 		"-config.file",
 		c.configFile,
+		"-limits.per-user-override-config",
+		c.overridesFile,
 	), flagset); err != nil {
 		return err
 	}
@@ -362,8 +431,11 @@ func (c *Component) run() error {
 	}()
 
 	c.cluster.waitGroup.Add(1)
+	c.wg.Add(1)
+
 	go func() {
 		defer c.cluster.waitGroup.Done()
+		defer c.wg.Done()
 		err := c.loki.Run(loki.RunOpts{})
 		if err != nil {
 			newErr := fmt.Errorf("error starting component %v: %w", c.name, err)
@@ -385,6 +457,7 @@ func (c *Component) run() error {
 func (c *Component) cleanup() (files []string, dirs []string) {
 	if c.loki != nil && c.loki.SignalHandler != nil {
 		c.loki.SignalHandler.Stop()
+		c.running = false
 	}
 	if c.configFile != "" {
 		files = append(files, c.configFile)
@@ -392,14 +465,43 @@ func (c *Component) cleanup() (files []string, dirs []string) {
 	if c.dataPath != "" {
 		dirs = append(dirs, c.dataPath)
 	}
-	if c.rulerWALPath != "" {
-		dirs = append(dirs, c.rulerWALPath)
-	}
-	if c.rulesPath != "" {
-		dirs = append(dirs, c.rulesPath)
-	}
 
 	return files, dirs
+}
+
+func (c *Component) Restart() error {
+	c.cleanup()
+	c.wg.Wait()
+	return c.run()
+}
+
+type runtimeConfigValues struct {
+	TenantLimits map[string]*validation.Limits `yaml:"overrides"`
+}
+
+func (c *Component) SetTenantLimits(tenant string, limits validation.Limits) error {
+	rcv := runtimeConfigValues{}
+	rcv.TenantLimits = c.loki.TenantLimits.AllByUserID()
+	if rcv.TenantLimits == nil {
+		rcv.TenantLimits = map[string]*validation.Limits{}
+	}
+	rcv.TenantLimits[tenant] = &limits
+
+	config, err := yaml.Marshal(rcv)
+	if err != nil {
+		return err
+	}
+
+	return os.WriteFile(c.overridesFile, config, 0777)
+}
+
+func (c *Component) GetTenantLimits(tenant string) validation.Limits {
+	limits := c.loki.TenantLimits.TenantLimits(tenant)
+	if limits == nil {
+		return c.loki.Cfg.LimitsConfig
+	}
+
+	return *limits
 }
 
 func NewRemoteWriteServer(handler *http.HandlerFunc) *httptest.Server {
